backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-socketio>=5.6.0",
    "strands-agents-tools>=0.2.19",
    "strands-agents[openai]>=1.22.0",
    "websocket-client>=1.9.0",
]
